requirements.txt
tensorflow
numpy
pandas
scikit-learn
matplotlib
seaborn
gradio
pillow
